Wrap chart updating in runOutsideAngular
Ran into a performance issue with this lib and Angular 9 Ivy (I don't know if this is related to Angular 9 or not, haven't tried to downgrade and compare).  When updating the chart's data, it sometimes would take a minute or two to re-draw the chart, during which time the app would eat 100% CPU of the Chrome process.  Profiling the process, it seemed to come from Timer calls, but not sure where they were originating from.

This change wraps the Chart update process in a `runOutsideAngular` call, similar to how the chart initialization process is done. This appears to greatly reduce the CPU usage when re-drawing the chart. I assume this has something to do with Angular trying to manage all of the SVG elements created when the chart is re-drawn, but not entirely sure.

There doesn't appear to be any negative side effects to doing this, as Angular doesn't need to monitor anything that the chart is building, but please correct me if I'm wrong.